feat: Add extra costs to invoice view (master)
Added a new section to display extra costs on the invoice view page.
This includes a KeyValueEntry to show the name and fee of each extra cost.
The fee is formatted using the IdrCurrency helper for proper display.
Also added a hint icon to the item price to indicate if it's included in the bill.